Move component and module schematics into a generate folder
