+++ src/generate/component/_files/__path__/__fileName__
