+++ src/generate/component/index_spec.ts
@@ -4,9 +4,9 @@ import { SchematicTestRunner } from '@angular-devkit/schematics/testing';
 import { getFileContent, createAppModule } from '@schematics/angular/utility/test';
 import { VirtualTree } from '@angular-devkit/schematics';
 
-import { createEmptyProject, toComponentClassName } from '../utils';
-import { DEFAULT_SHARED_EXTENSIONS } from '../generate/utils';
-import { isInComponentMetadata } from '../test-utils';
+import { createEmptyProject, toComponentClassName } from '../../utils';
+import { DEFAULT_SHARED_EXTENSIONS } from '../utils';
+import { isInComponentMetadata } from '../../test-utils';
 import { Schema as ComponentOptions } from './schema';
 
 describe('Component Schematic', () => {
