+++ src/generate/component/schema.d.ts
