+++ src/generate/component/schema.json
