+++ src/generate/module/index_spec.ts
@@ -5,9 +5,9 @@ import { SchematicTestRunner } from '@angular-devkit/schematics/testing';
 import { getFileContent, createAppModule } from '@schematics/angular/utility/test';
 
 import { Schema as ModuleOptions } from './schema';
-import { createEmptyProject, toNgModuleClassName } from '../utils';
-import { DEFAULT_SHARED_EXTENSIONS } from '../generate/utils';
-import { isInModuleMetadata } from '../test-utils';
+import { createEmptyProject, toNgModuleClassName } from '../../utils';
+import { DEFAULT_SHARED_EXTENSIONS } from '../utils';
+import { isInModuleMetadata } from '../../test-utils';
 
 describe('Module Schematic', () => {
   const path = 'app';
